Moved CfgFunctions
